Mostly working
Everything is working.  Ran into an error where the first step would
skip because of an '\n' left in stdin so I built a small buffer clear in
debug monitor model.  Unfortunately the buffer clear halts the program
if there is no '\n' to clear.  This means it requires two enter presses
to move foward during step.  Not sure how to resolve that.